- enhanced update availability api endpoint by supporing range based slots and interval setting field
- minor cleanup moved generate slots method into the availability mixin
